fast_api_scraping/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
APScheduler==3.11.0
fastapi==0.115.6
uvicorn==0.30.6
confluent-kafka==2.7.0
python-dotenv==1.0.1